Use uint8_t in msg/multi/sync and fix MSVC warnings (#6856)
* Fix MSVC warnings in msg.cpp

* Move TBuffer to multi.cpp

* Fix MSVC warnings in multi.cpp

* SNet* change databytes to size_t

* msg/muli/sync: Change playerId to uint8_t

+++ Source/dvlnet/abstract_net.h
@@ -18,10 +18,10 @@ class abstract_net {
 public:
 	virtual int create(std::string addrstr) = 0;
 	virtual int join(std::string addrstr) = 0;
-	virtual bool SNetReceiveMessage(uint8_t *sender, void **data, uint32_t *size) = 0;
-	virtual bool SNetSendMessage(int dest, void *data, unsigned int size) = 0;
+	virtual bool SNetReceiveMessage(uint8_t *sender, void **data, size_t *size) = 0;
+	virtual bool SNetSendMessage(uint8_t dest, void *data, size_t size) = 0;
 	virtual bool SNetReceiveTurns(char **data, size_t *size, uint32_t *status) = 0;
-	virtual bool SNetSendTurn(char *data, unsigned int size) = 0;
+	virtual bool SNetSendTurn(char *data, size_t size) = 0;
 	virtual void SNetGetProviderCaps(struct _SNETCAPS *caps) = 0;
 	virtual bool SNetRegisterEventHandler(event_type evtype, SEVTHANDLER func) = 0;
 	virtual bool SNetUnregisterEventHandler(event_type evtype) = 0;
