+++ Source/dvlnet/loopback.h
@@ -19,10 +19,10 @@ public:
 
 	int create(std::string addrstr) override;
 	int join(std::string addrstr) override;
-	bool SNetReceiveMessage(uint8_t *sender, void **data, uint32_t *size) override;
-	bool SNetSendMessage(int dest, void *data, unsigned int size) override;
+	bool SNetReceiveMessage(uint8_t *sender, void **data, size_t *size) override;
+	bool SNetSendMessage(uint8_t dest, void *data, size_t size) override;
 	bool SNetReceiveTurns(char **data, size_t *size, uint32_t *status) override;
-	bool SNetSendTurn(char *data, unsigned int size) override;
+	bool SNetSendTurn(char *data, size_t size) override;
 	void SNetGetProviderCaps(struct _SNETCAPS *caps) override;
 	bool SNetRegisterEventHandler(event_type evtype, SEVTHANDLER func) override;
 	bool SNetUnregisterEventHandler(event_type evtype) override;
